input: Add is_keyboard to PlayerControl to detect the movement source
When using a joystick, movements are (and need to be) converted to
virtual keypresses. This causes the player movement code to ignore
the analog joystick value and makes the player always walk at full
speed.

This patch adds a boolean value to PlayerControl, which allows the
movement code to detect whether the input came from the keyboard
or from the joystick, and adjust the walking speed accordingly.